requirements.txt
# Core dependencies
google-generativeai>=0.3.0
PyPDF2>=3.0.0
radon>=6.0.1
pylint>=3.0.0
astor>=0.8.1

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

streamlit>=1.28.0
plotly>=5.18.0
markdown2>=2.4.10
reportlab>=4.0.0